requirements.txt
asgiref==3.6.0
Django==3.2
Faker==16.6.0
flake8==4.0.0
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
django-split-settings
django-debug-toolbar==3.2.4
gunicorn==20.0.4


